+++ dart/geocore/lib/src/base/point_wrapper.dart
@@ -56,10 +56,17 @@ class PointWrapper<T extends Point<C>, C extends num> extends Point<C>
   C get m => point.m;
 
   @override
-  Point newWith({num x = 0.0, num y = 0.0, num? z, num? m}) =>
-      point.newWith(x: x, y: y, z: z, m: m);
+  T copyWith({num? x, num? y, num? z, num? m}) =>
+      point.copyWith(x: x, y: y, z: z, m: m) as T;
+      
+  @override
+  T newWith({num x = 0.0, num y = 0.0, num? z, num? m}) =>
+      point.newWith(x: x, y: y, z: z, m: m) as T;
 
   @override
-  Point newFrom(Iterable<num> coords, {int? offset, int? length}) =>
-      point.newFrom(coords, offset: offset, length: length);
+  T newFrom(Iterable<num> coords, {int? offset, int? length}) =>
+      point.newFrom(coords, offset: offset, length: length) as T;
+
+  @override
+  T project(TransformPoint transform) => point.project(transform) as T;
 }