+++ dart/geocore/lib/src/feature/feature.dart
@@ -67,6 +67,13 @@ abstract class Feature<T extends Geometry> implements Bounded {
 
   /// An optional geometry for this feature.
   Geometry? get geometry;
+
+  /// Returns a new feature with geometry projected using [transform].
+  /// 
+  /// Projects [geometry] of this feature. Other members, [id] and [properties],
+  /// are set without modifications to a new feature object. 
+  @override
+  Feature<T> project(TransformPoint transform);
 }
 
 /// Private implementation of [Feature].
@@ -97,6 +104,14 @@ class _FeatureBase<T extends Geometry>
 
   @override
   Bounds get bounds => _featureBounds ?? geometry?.bounds ?? Bounds.empty();
+
+  @override
+  Feature<T> project(TransformPoint transform) => _FeatureBase(
+        id: id,
+        properties: properties,
+        geometry: geometry?.project(transform) as T?,
+        bounds: _featureBounds?.project(transform),
+      );
 }
 
 /// A feature collection with a series of features.
@@ -115,6 +130,10 @@ abstract class FeatureCollection<T extends Feature> extends Bounded {
 
   /// All the [features] for this collection.
   BoundedSeries<T> get features;
+
+  /// Returns a new collection with features projected using [transform].
+  @override
+  FeatureCollection<T> project(TransformPoint transform);
 }
 
 /// Private implementation of [FeatureCollection].
@@ -135,4 +154,11 @@ class _FeatureCollectionBase<T extends Feature> extends FeatureCollection<T>
 
   @override
   Bounds get bounds => _collectionBounds ?? features.bounds;
+
+  @override
+  FeatureCollection<T> project(TransformPoint transform) =>
+      _FeatureCollectionBase(
+        features: features.project(transform, lazy: false),
+        bounds: _collectionBounds?.project(transform),
+      );
 }